Version1 of TreWell
The current version of the TreWell website includes all the essential features such as account creation or login to track finances, displaying the main balance with options to add, deduct, or reset funds, managing short-term financial categories by specifying the name, description, target amount, and date, with the ability to deposit and withdraw from each category. The site also allows viewing deleted categories and restoring or permanently deleting them. It features an attractive interface, pop-up messages for action confirmations, and motivational messages. All data is saved using local storage, making this version a practical and stable release of the project so far.